Move the function go retrieve test file to pylint.testutil.functional
