+++ pylint/testutils/functional/__init__.py
@@ -3,10 +3,16 @@
 
 __all__ = [
     "FunctionalTestFile",
+    "REASONABLY_DISPLAYABLE_VERTICALLY",
+    "get_functional_test_files_from_directory",
     "NoFileError",
     "parse_python_version",
 ]
 
+from pylint.testutils.functional.find_functional_tests import (
+    REASONABLY_DISPLAYABLE_VERTICALLY,
+    get_functional_test_files_from_directory,
+)
 from pylint.testutils.functional.test_file import (
     FunctionalTestFile,
     NoFileError,
